app/requirements.txt
amqp==5.1.1
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.3.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-PyMongo==2.3.0
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
nanoid==2.0.0
numpy==1.24.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
Pillow==9.5.0
prompt-toolkit==3.0.38
pydantic==1.10.2
pymongo==3.13.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.30.0
six==1.16.0
SQLAlchemy==1.4.44
SQLAlchemy-Utils==0.38.3
typing_extensions==4.5.0
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.4